backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
